Backend/user_domain/sessionService/readSession/requirements.txt
flask
redis
python-dotenv
requests
flasgger